Extract functions to create BSON instances
