+++ src/BSON/Binary.c
@@ -354,3 +354,15 @@ void php_phongo_binary_init_ce(INIT_FUNC_ARGS)
 	php_phongo_handler_binary.free_obj       = php_phongo_binary_free_object;
 	php_phongo_handler_binary.offset         = XtOffsetOf(php_phongo_binary_t, std);
 }
+
+void php_phongo_bson_new_binary(zval* object, const char* data, size_t data_len, bson_subtype_t type)
+{
+	php_phongo_binary_t* intern;
+
+	object_init_ex(object, php_phongo_binary_ce);
+
+	intern           = Z_BINARY_OBJ_P(object);
+	intern->data     = estrndup(data, data_len);
+	intern->data_len = data_len;
+	intern->type     = (uint8_t) type;
+}
