+++ src/BSON/Decimal128.c
@@ -289,3 +289,14 @@ void php_phongo_decimal128_init_ce(INIT_FUNC_ARGS)
 	php_phongo_handler_decimal128.free_obj       = php_phongo_decimal128_free_object;
 	php_phongo_handler_decimal128.offset         = XtOffsetOf(php_phongo_decimal128_t, std);
 }
+
+void php_phongo_bson_new_decimal128(zval* object, const bson_decimal128_t* decimal)
+{
+	php_phongo_decimal128_t* intern;
+
+	object_init_ex(object, php_phongo_decimal128_ce);
+
+	intern = Z_DECIMAL128_OBJ_P(object);
+	memcpy(&intern->decimal, decimal, sizeof(bson_decimal128_t));
+	intern->initialized = true;
+}
