+++ src/MongoDB/Monitoring/CommandSucceededEvent.c
@@ -23,6 +23,7 @@
 #include "phongo_bson_encode.h"
 #include "phongo_error.h"
 
+#include "BSON/Int64.h"
 #include "BSON/ObjectId.h"
 #include "MongoDB/Server.h"
 #include "CommandSucceededEvent_arginfo.h"
@@ -127,7 +128,7 @@ static PHP_METHOD(MongoDB_Driver_Monitoring_CommandSucceededEvent, getServiceId)
 		RETURN_NULL();
 	}
 
-	phongo_objectid_init(return_value, &intern->service_id);
+	php_phongo_bson_new_objectid(return_value, &intern->service_id);
 }
 
 /* Returns the event's server connection ID */
@@ -230,7 +231,7 @@ static HashTable* php_phongo_commandsucceededevent_get_debug_info(phongo_compat_
 	if (intern->has_service_id) {
 		zval service_id;
 
-		phongo_objectid_init(&service_id, &intern->service_id);
+		php_phongo_bson_new_objectid(&service_id, &intern->service_id);
 		ADD_ASSOC_ZVAL_EX(&retval, "serviceId", &service_id);
 	} else {
 		ADD_ASSOC_NULL_EX(&retval, "serviceId");
