+++ src/MongoDB/Monitoring/ServerClosedEvent.c
@@ -55,7 +55,7 @@ static PHP_METHOD(MongoDB_Driver_Monitoring_ServerClosedEvent, getTopologyId)
 
 	PHONGO_PARSE_PARAMETERS_NONE();
 
-	phongo_objectid_init(return_value, &intern->topology_id);
+	php_phongo_bson_new_objectid(return_value, &intern->topology_id);
 }
 
 /* MongoDB\Driver\Monitoring\ServerClosedEvent object handlers */
@@ -94,7 +94,7 @@ static HashTable* php_phongo_serverclosedevent_get_debug_info(phongo_compat_obje
 
 	{
 		zval topology_id;
-		phongo_objectid_init(&topology_id, &intern->topology_id);
+		php_phongo_bson_new_objectid(&topology_id, &intern->topology_id);
 		ADD_ASSOC_ZVAL_EX(&retval, "topologyId", &topology_id);
 	}
